refactor(Whole_project): Created game-state files to streamline conditions.
Lots of minor refactors within game and entities to add/change a game-state depending on situation. Game loop will go on as long as game-state enum is set to "running". This will allow us to call it from anywhere to change state without hard coding it.

Added a couple TODOS for TT