build(ItemRepository): Updated itemRepo to have a Database
A database with all generated items that the game will refer to when spawning in items.
This ensures that players and npc's will share the same reference number and allow it to stack with itself when sold/bought.

Currently only able to get by string ID which means that it can be spelt wrong by developer.