feat(Combat): Combat printing is more streamlined✨✨
Adjusted wordings to be more straight to the point.
Initiative only prints on the 1st turn to point out who's starting.
Not needed to be shown due to it being a back and forth afterwards.
Combat now prints 1 page per action and alternates between player/enemy.
Added spacing and turn end to create a sense of "loop"